refactor: Lotto클래스의 상수 제거 후 LottoNumber 정적상수 활용
