+++ src/main/java/lotto/constant/ExceptionMessages.java
@@ -1,6 +1,6 @@
 package lotto.constant;
 
-public class MessageConstant {
+public class ExceptionMessages {
     public static final String INVALID_DUPLICATED_LOTTO_NUMBER = "로또에 중복된 숫자가 존재할 수 없습니다.";
     public static final String INVALID_NUMBER_RANGE = "로또 숫자는 1이상 45이하의 범위를 가져야 합니다.";
     public static final String INVALID_LOTTO_SIZE = "로또 숫자는 6개여야 합니다.";
