+++ src/main/java/lotto/constant/LottoGrade.java
@@ -1,6 +1,6 @@
 package lotto.constant;
 
-import static lotto.constant.MessageConstant.INVALID_MATCH_COUNT_RANGE;
+import static lotto.constant.ExceptionMessages.INVALID_MATCH_COUNT_RANGE;
 
 import java.util.Arrays;
